+++ sample/Farkle.Samples.CSharp/JSON.cs
@@ -8,7 +8,6 @@ using Microsoft.FSharp.Collections;
 using System;
 using System.Globalization;
 using System.Text;
-using Farkle.CSharp;
 using Farkle.Builder;
 using static Farkle.Builder.Regex;